+++ cypress/integration/e2e/menus/emotion.spec.ts
@@ -1,3 +1,6 @@
+import menus from '../../../../src/config/menus'
+// 按钮位置
+const pos = menus.menus.indexOf('emoticon')
 describe('表情', () => {
     beforeEach(() => {
         cy.visit('/examples/index.html')
@@ -13,7 +16,7 @@ describe('表情', () => {
         cy.get('@Editable').type(text)
         cy.get('@Editable').contains(text)
 
-        cy.getByClass('toolbar').children().eq(15).as('emotionMenu').click()
+        cy.getByClass('toolbar').children().eq(pos).as('emotionMenu').click()
 
         cy.get('@emotionMenu').find('.w-e-panel-container').as('Panel').should('be.visible')
         cy.get('@Panel').find('.w-e-panel-tab-title').children().should('have.length', 4)
@@ -29,7 +32,7 @@ describe('表情', () => {
         cy.get('@Editable').type(text)
         cy.get('@Editable').contains(text)
 
-        cy.getByClass('toolbar').children().eq(15).as('emotionMenu').click()
+        cy.getByClass('toolbar').children().eq(pos).as('emotionMenu').click()
 
         cy.get('@emotionMenu').find('.w-e-panel-container').as('Panel').should('be.visible')
         cy.get('@Panel').find('.w-e-panel-close').click()
@@ -41,21 +44,23 @@ describe('表情', () => {
         cy.get('@Editable').type(text)
         cy.get('@Editable').contains(text)
 
-        cy.getByClass('toolbar').children().eq(15).as('emotionMenu').click()
+        cy.getByClass('toolbar').children().eq(pos).as('emotionMenu').click()
 
         cy.get('@emotionMenu').find('.w-e-panel-container').as('Panel').should('be.visible')
         cy.get('@Panel').find('.w-e-panel-tab-title').children().should('have.length', 4)
+        cy.get('@Panel').find('.w-e-panel-tab-title').children().eq(2).click()
+
         cy.get('@Panel')
             .find('.w-e-panel-tab-title')
             .children()
-            .eq(0)
+            .eq(2)
             .should('have.class', 'w-e-active')
-            .and('contain.text', '默认')
+            .and('contain.text', 'emoji')
 
         cy.get('@Panel')
             .find('.w-e-panel-tab-content')
             .children()
-            .eq(0)
+            .eq(2)
             .children()
             .as('emotionList')
 
@@ -64,10 +69,9 @@ describe('表情', () => {
             .as('emotion')
             .click()
             .then($el => {
-                const img = $el.find('img')
-                const src = (img.get(0) as HTMLImageElement).src
+                const emotionValue = $el.get(0).innerText
 
-                cy.get('@Editable').find('img').should('have.attr', 'src', src)
+                cy.get('@Editable').should('contain.text', emotionValue)
             })
     })
 })