+++ cypress/integration/e2e/menus/insertVideo.spec.ts
@@ -1,3 +1,6 @@
+import menus from '../../../../src/config/menus'
+// 按钮位置
+const pos = menus.menus.indexOf('video')
 describe('插入视频', () => {
     beforeEach(() => {
         cy.visit('/examples/index.html')
@@ -7,11 +10,10 @@ describe('插入视频', () => {
         cy.get('@Editable').clear()
     })
 
-    const videoUrl =
-        'https://player.vimeo.com/external/288190258.sd.mp4?s=8bbc2d25f23cf412a99bd2dbdfa08688fd973ce8&profile_id=164&oauth2_token_id=57447761'
+    const videoUrl = 'https://www.bilibili.com/video/BV14p4y1v776/'
 
     it('点击菜单打开插入视频的面板', () => {
-        cy.getByClass('toolbar').children().eq(17).as('videoMenu').click()
+        cy.getByClass('toolbar').children().eq(pos).as('videoMenu').click()
 
         cy.get('@videoMenu').find('.w-e-panel-container').as('Panel').should('be.visible')
         cy.get('@Panel').find('.w-e-panel-tab-title').contains('插入视频')
@@ -21,7 +23,7 @@ describe('插入视频', () => {
     })
 
     it('点击插入视频的面板的关闭按钮可以收起面板', () => {
-        cy.getByClass('toolbar').children().eq(17).as('videoMenu').click()
+        cy.getByClass('toolbar').children().eq(pos).as('videoMenu').click()
 
         cy.get('@videoMenu').find('.w-e-panel-container').as('Panel').should('be.visible')
         cy.get('@Panel').find('.w-e-panel-tab-title').contains('插入视频')
@@ -32,16 +34,15 @@ describe('插入视频', () => {
     it('插入网络视频', () => {
         cy.get('@Editable').find('video').should('not.exist')
 
-        cy.getByClass('toolbar').children().eq(17).as('imgMenu').click()
-        const videoEl = `<video src="${videoUrl}" controls></video>`
+        cy.getByClass('toolbar').children().eq(pos).as('imgMenu').click()
+        const videoEl = `<iframe src="${videoUrl}" controls></iframe>`
         cy.get('@imgMenu').find('.w-e-panel-container').as('Panel').find('input').type(videoEl)
         cy.get('@Panel').find('.w-e-button-container button').click()
 
-        cy.get('@Editable')
-            .find('video')
+        cy.get('.w-e-text-container iframe', { timeout: 10000 })
             .should('be.visible')
             .then($video => {
-                const video = $video.get(0)
+                const video = $video.get(0) as HTMLVideoElement
                 expect(video.src).to.eq(videoUrl)
             })
     })