+++ src/config/index.ts
@@ -12,6 +12,7 @@ import imageConfig, { UploadImageHooksType } from './image'
 import textConfig from './text'
 import langConfig from './lang'
 import historyConfig from './history'
+import videoConfig, { UploadVideoHooksType } from './video'
 
 // 字典类型
 export type DicType = {
@@ -24,6 +25,7 @@ export type ConfigType = {
     languageType: string[]
     languageTab: string
     menus: string[]
+    excludeMenus: string[]
     fontNames: string[]
     lineHeights: string[]
     indentation: IndentationType
@@ -45,6 +47,7 @@ export type ConfigType = {
     zIndexFullScreen: number
     showFullScreen: boolean
     showLinkImg: boolean
+    uploadImgAccept: string[]
     uploadImgServer: string
     uploadImgShowBase64: boolean
     uploadImgMaxSize: number
@@ -57,7 +60,9 @@ export type ConfigType = {
     uploadImgTimeout: number
     withCredentials: boolean
     customUploadImg: Function | null
-    customAlert: Function | null
+    customAlert: Function
+
+    onCatalogChange: Function | null
 
     lang: string
     languages: typeof langConfig
@@ -68,6 +73,23 @@ export type ConfigType = {
     historyMaxSize: number
 
     focus: boolean
+
+    onlineVideoCheck: Function
+    onlineVideoCallback: Function
+
+    showLinkVideo: Boolean
+    uploadVideoAccept: string[]
+    uploadVideoServer: string
+    uploadVideoMaxSize: number
+    uploadVideoName: string
+    uploadVideoParams: DicType
+    uploadVideoParamsWithUrl: boolean
+    uploadVideoHeaders: DicType
+    uploadVideoHooks: UploadVideoHooksType
+    uploadVideoTimeout: number
+    withVideoCredentials: boolean
+    customUploadVideo: Function | null
+    customInsertVideo: Function | null
 }
 
 export type Resource = {
@@ -96,6 +118,7 @@ const defaultConfig = Object.assign(
     textConfig,
     langConfig,
     historyConfig,
+    videoConfig,
     //链接校验的配置函数
     {
         linkCheck: function (text: string, link: string): string | boolean {